Modification to the existing trisyclcc script to mimic the syclcc script except in this case instead of versions it cycles through compile modes..
Not completely sold on this idea yet, however triSYCL has a lot of different ways to compile and they all require slightly different options. You can checkout triSYCL/tests/Makefile to get an idea.

Currently only the OpenMP script is working and it's what I'm currently using to test SLAMBench support for triSYCL, the other scripts showcase other compile targets that work for the moment and can have scripts implemented in the future.