Large restructuring and renaming commit, minor modifications to triSYCL script to align with ComputeCpp script naming conventions
This is an experimental direction/push for now, would be interested in 
feedback on this. In essence I thought it might be a good idea to rename 
the ComputeCpp related syclcc script to computecppcc (as ComputeCPP is a 
SYCL implementation) and restructure the directories a little to avoid 
bloat.

The renaming however does mean that syclcc is more of an overarching 
project name than the specific name of the scripts... still a good name 
though! It may also be possible to implement a top level script called 
syclcc that will forward to the trisyclcc or computecppcc scripts 
(perhaps even future implementations scripts?).

I have split the scripts into seperate directories, one for ComputeCpp 
scripts(computecpp) and the other for triSYCL(trisycl) scripts.

I have renamed the syclcc-ComputeCpp* scripts to 
computecppcc-ComputeCpp* and the syclcc script has been renamed to 
computecppcc.

The trisycl scripts were not renamed and continue to have trisyclcc as 
the main script name. Some minor modificaitons like changing the name of 
TRISYCL to TRISYCL_DIR were implemented.

I have also modified the readme accordingly to reflect the new triSYCL 
scripts and subsequent renaming of syclcc.