Mechanism for mapping ComputeCpp versions to 'generic' scripts
I have added a version -> script map to aid selection of generic scripts
and future scripts. It's also easily modifiable for users if they wish
to add new scripts. This replaces the introspection of the ComputeCpp
version name, which will get problematic as the version name changes
frequently (for example recently in 2.0.0)